feat: support fractional calibration values

+++ Models/CalibPoint.cs
@@ -5,7 +5,7 @@ public class CalibPoint
     public int     Id       { get; set; }
     public int     Channel  { get; set; }  // 0 = CH0, 1 = CH1
     public int     AdcCode  { get; set; }
-    public int     CalibrationValue { get; set; }
+    public decimal CalibrationValue { get; set; }
     public decimal Mass     { get; set; }
     public bool    IsActive { get; set; } = true;
     public DateTime  CreatedAt { get; set; }
